refactor(frontend): move 2 files into one
