+++ apps/frontend/app/lib/utilities.server.ts
@@ -1,14 +1,21 @@
 import { parseWithZod } from "@conform-to/zod";
+import { $path } from "@ignisda/remix-routes";
 import {
 	json,
+	redirect,
 	unstable_composeUploadHandlers,
 	unstable_createMemoryUploadHandler,
 } from "@remix-run/node";
 import {
+	type CoreDetails,
+	CoreEnabledFeaturesDocument,
 	GetPresignedS3UrlDocument,
 	PresignedPutS3UrlDocument,
+	type UserCollectionsListQuery,
 	type UserLot,
+	type UserPreferences,
 } from "@ryot/generated/graphql/backend/graphql";
+import { withQuery, withoutHost } from "ufo";
 import { type ZodTypeAny, type output, z } from "zod";
 import { API_URL, gqlClient } from "./api.server";
 import {
@@ -18,6 +25,7 @@ import {
 	userDetailsCookie,
 	userPreferencesCookie,
 } from "./cookies.server";
+import { redirectToQueryParam } from "./generals";
 
 export const expectedEnvironmentVariables = z.object({
 	DISABLE_TELEMETRY: z
@@ -179,3 +187,51 @@ export const s3FileUploader = (prefix: string) =>
 		}
 		return undefined;
 	}, unstable_createMemoryUploadHandler());
+
+export const getCoreEnabledFeatures = async () => {
+	const { coreEnabledFeatures } = await gqlClient.request(
+		CoreEnabledFeaturesDocument,
+	);
+	return coreEnabledFeatures;
+};
+
+export const getCoreDetails = async (request: Request) => {
+	const details = await coreDetailsCookie.parse(
+		request.headers.get("cookie") || "",
+	);
+	redirectIfDetailNotPresent(request, details);
+	return details as CoreDetails;
+};
+
+export const getUserPreferences = async (request: Request) => {
+	const prefs = await userPreferencesCookie.parse(
+		request.headers.get("cookie") || "",
+	);
+	redirectIfDetailNotPresent(request, prefs);
+	return prefs as UserPreferences;
+};
+
+export const getUserDetails = async (request: Request) => {
+	const details = await userDetailsCookie.parse(
+		request.headers.get("cookie") || "",
+	);
+	redirectIfDetailNotPresent(request, details);
+	return details as ApplicationUser;
+};
+
+export const getUserCollectionsList = async (request: Request) => {
+	const list = await userCollectionsListCookie.parse(
+		request.headers.get("cookie") || "",
+	);
+	redirectIfDetailNotPresent(request, list);
+	return list as UserCollectionsListQuery["userCollectionsList"];
+};
+
+const redirectIfDetailNotPresent = (request: Request, detail: unknown) => {
+	if (!detail)
+		throw redirect(
+			withQuery($path("/actions"), {
+				[redirectToQueryParam]: withoutHost(request.url),
+			}),
+		);
+};
