+++ apps/frontend/app/routes/_dashboard.collections.list.public.tsx
@@ -24,8 +24,8 @@ import {
 	DebouncedSearchInput,
 } from "~/components/common";
 import { gqlClient } from "~/lib/api.server";
-import { getCoreDetails } from "~/lib/graphql.server";
 import { useGetMantineColor, useSearchParam } from "~/lib/hooks";
+import { getCoreDetails } from "~/lib/utilities.server";
 
 const searchParamsSchema = z.object({
 	page: zx.IntAsString.optional(),
