+++ apps/frontend/app/routes/_dashboard.fitness.measurements.list.tsx
@@ -50,10 +50,9 @@ import { zx } from "zodix";
 import { getAuthorizationHeader, gqlClient } from "~/lib/api.server";
 import events from "~/lib/events";
 import { ApplicationKey, dayjsLib, redirectToQueryParam } from "~/lib/generals";
-import { getUserPreferences } from "~/lib/graphql.server";
 import { useSearchParam } from "~/lib/hooks";
 import { createToastHeaders } from "~/lib/toast.server";
-import { processSubmission } from "~/lib/utilities.server";
+import { getUserPreferences, processSubmission } from "~/lib/utilities.server";
 
 enum TimeSpan {
 	Last7Days = "Last 7 days",
