+++ apps/frontend/app/routes/_dashboard.fitness.workouts.current.tsx
@@ -93,12 +93,12 @@ import { DisplayExerciseStats } from "~/components/fitness";
 import { getAuthorizationHeader, gqlClient } from "~/lib/api.server";
 import events from "~/lib/events";
 import { ApplicationKey, dayjsLib, getSetColor } from "~/lib/generals";
+import { createToastHeaders, redirectWithToast } from "~/lib/toast.server";
 import {
 	getCoreDetails,
 	getCoreEnabledFeatures,
 	getUserPreferences,
-} from "~/lib/graphql.server";
-import { createToastHeaders, redirectWithToast } from "~/lib/toast.server";
+} from "~/lib/utilities.server";
 import {
 	type Exercise,
 	type ExerciseSet,
