+++ apps/frontend/app/routes/_dashboard.fitness.workouts.list.tsx
@@ -41,8 +41,8 @@ import {
 import { getSetStatisticsTextToDisplay } from "~/components/fitness";
 import { getAuthorizationHeader, gqlClient } from "~/lib/api.server";
 import { dayjsLib } from "~/lib/generals";
-import { getCoreDetails, getUserPreferences } from "~/lib/graphql.server";
 import { getWorkoutStarter, useSearchParam } from "~/lib/hooks";
+import { getCoreDetails, getUserPreferences } from "~/lib/utilities.server";
 import { getDefaultWorkout } from "~/lib/workout";
 
 const searchParamsSchema = z.object({
