+++ apps/frontend/app/routes/_dashboard.media.create.tsx
@@ -32,8 +32,11 @@ import { camelCase, changeCase } from "@ryot/ts-utils";
 import { IconCalendar, IconPhoto, IconVideo } from "@tabler/icons-react";
 import { z } from "zod";
 import { getAuthorizationHeader, gqlClient } from "~/lib/api.server";
-import { getCoreEnabledFeatures } from "~/lib/graphql.server";
-import { processSubmission, s3FileUploader } from "~/lib/utilities.server";
+import {
+	getCoreEnabledFeatures,
+	processSubmission,
+	s3FileUploader,
+} from "~/lib/utilities.server";
 
 export const loader = async (_args: LoaderFunctionArgs) => {
 	const [coreEnabledFeatures] = await Promise.all([getCoreEnabledFeatures()]);
