+++ apps/frontend/app/routes/_dashboard.settings.profile.tsx
@@ -18,8 +18,8 @@ import { useRef } from "react";
 import { z } from "zod";
 import { confirmWrapper } from "~/components/confirmation";
 import { getAuthorizationHeader, gqlClient } from "~/lib/api.server";
-import { getUserDetails } from "~/lib/graphql.server";
 import { createToastHeaders } from "~/lib/toast.server";
+import { getUserDetails } from "~/lib/utilities.server";
 import { processSubmission } from "~/lib/utilities.server";
 
 export const loader = async ({ request }: LoaderFunctionArgs) => {
