+++ apps/frontend/app/routes/_dashboard.tsx
@@ -52,8 +52,11 @@ import { HiddenLocationInput } from "~/components/common";
 import { redirectIfNotAuthenticated } from "~/lib/api.server";
 import { colorSchemeCookie } from "~/lib/cookies.server";
 import { ApplicationKey, getLot } from "~/lib/generals";
-import { getCoreDetails, getUserPreferences } from "~/lib/graphql.server";
-import { expectedEnvironmentVariables } from "~/lib/utilities.server";
+import {
+	expectedEnvironmentVariables,
+	getCoreDetails,
+	getUserPreferences,
+} from "~/lib/utilities.server";
 import classes from "~/styles/dashboard.module.css";
 
 export const loader = async ({ request }: LoaderFunctionArgs) => {
